fix(mapi): updated the approach to file upload to not presigned urls
refs. metriport/metriport#960
