+++ packages/utils/src/document-upload.ts
@@ -2,10 +2,9 @@ import * as dotenv from "dotenv";
 dotenv.config();
 // keep that ^ on top
 import { DocumentReference } from "@medplum/fhirtypes";
+import fs from "fs/promises";
 import { MetriportMedicalApi } from "@metriport/api-sdk";
 import { getEnvVarOrFail } from "@metriport/core/util/env-var";
-import FormData from "form-data";
-import { createReadStream } from "fs";
 
 const apiKey = getEnvVarOrFail("API_KEY");
 const apiUrl = getEnvVarOrFail("API_URL");
@@ -15,9 +14,17 @@ const metriportApi = new MetriportMedicalApi(apiKey, {
   baseAddress: apiUrl,
 });
 
+async function readFileAsync(filePath: string) {
+  try {
+    return await fs.readFile(filePath, "utf8");
+  } catch (err) {
+    console.error(err);
+  }
+}
+
 async function main() {
-  const documentReference: Partial<DocumentReference> = {
-    description: "Third degree wrist burn treatment",
+  const docRef: Partial<DocumentReference> = {
+    description: "Second degree ankle burn treatment",
     type: {
       text: "Burn management Hospital Progress note",
       coding: [
@@ -36,27 +43,10 @@ async function main() {
       },
     },
   };
-  const presignedUploadUrl = await metriportApi.getDocumentUploadUrl(patientId, documentReference);
+  const fileContent = await readFileAsync("./src/shorter_example.xml");
+  if (!fileContent) throw new Error("File content is empty");
 
-  console.log("Presigned URL:\n", JSON.stringify(presignedUploadUrl, null, 2));
-  const form = new FormData();
-  Object.entries(presignedUploadUrl.fields).forEach(([field, value]) => {
-    form.append(field, value);
-  });
-  const contentType = "application/xml";
-  form.append("Content-Type", contentType);
-  form.append("file", createReadStream("./src/shorter_example.xml"));
-
-  console.log("Uploading the file now...");
-  form.submit(presignedUploadUrl.url, (err, response) => {
-    if (err) {
-      console.log(`ERROR: `, err);
-      return;
-    }
-    console.log(`RESPONSE: `, response.statusCode);
-  });
-
-  console.log("To post-process the file, use the following script: document-upload-post-process");
+  await metriportApi.uploadDocument(patientId, docRef, fileContent);
 }
 
 main();
